requirements.txt
Flask>=2.0
flask-cors>=3.0
requests>=2.28
cryptography>=40.0
pycryptodome>=3.18
imapclient>=2.3
bcrypt>=4.0
flasgger>=0.9
